+++ tests/Factories/TaskFactory.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Sfneal\Helpers\Time\Tests\Factories;
+
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Sfneal\Helpers\Time\Tests\Models\Task;
+
+class TaskFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Task::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition(): array
+    {
+        return [
+            'total_duration' => $this->faker->randomFloat(2, 1, 240),
+        ];
+    }
+}